requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
pydub>=0.25.0
PyYAML>=6.0
pandas>=2.0.0
tqdm>=4.65.0
yt-dlp>=2023.1.0
